fix: move test files to __test__
