+++ package.json
@@ -20,7 +20,7 @@
     "start:fullstack": "cross-env REACT_APP_WHOAMI=fullstack react-scripts start",
     "start:frontend": "cross-env REACT_APP_WHOAMI=frontend react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test ./__test__",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
